Add hostname to category list for groups
The group that gets applied must match the hostname,
but if there are host-specific files they are currently
ignored.

Take the following `filelist`:
```
laptop=vim,shell

.vimrc:vim
.bash_profile:shell
.gitconfig:laptop
```
On the computer with hostname `laptop`, the files
`.vimrc` and `.bash_profile` would be handled as
intended, but `.gitconfig` would be ignored unless
you modify the group line to look like this:
```
laptop=vim,shell,laptop
```

In order to allow for different versions of files on
different hosts, when the group name matches the current
hostname add that to the list of categories for the group.